requirements.txt
aiogram>=3.4
sqlalchemy>=2.0
aiosqlite
openai>=1.0
pydub
python-dotenv
